use integer attribute for mesh index

+++ src/renderer/mergeMeshesToGeometry.js
@@ -52,7 +52,7 @@ function mergeGeometry(geometryAndMaterialIndex, vertexCount, indexCount) {
   const positionAttrib = new BufferAttribute(new Float32Array(3 * vertexCount), 3, false);
   const normalAttrib = new BufferAttribute(new Float32Array(3 * vertexCount), 3, false);
   const uvAttrib = new BufferAttribute(new Float32Array(2 * vertexCount), 2, false);
-  const materialMeshIndexAttrib = new BufferAttribute(new Float32Array(2 * vertexCount), 2, false);
+  const materialMeshIndexAttrib = new BufferAttribute(new Int32Array(2 * vertexCount), 2, false);
   const indexAttrib = new BufferAttribute(new Uint32Array(indexCount), 1, false);
 
   const mergedGeometry = new BufferGeometry();
